+++ src/app/static/src/tests/components/modelOutput/modelOutput.test.ts
@@ -7,6 +7,7 @@ import {
     mockComparisonPlotResponse,
     mockModelCalibrateState,
     mockShapeResponse,
+    mockError
 } from "../../mocks";
 import {mutations as modelOutputMutations} from "../../../app/store/modelOutput/mutations";
 import {mutations as plottingSelectionMutations} from "../../../app/store/plottingSelections/mutations";
@@ -21,10 +22,11 @@ import {expectTranslated} from "../../testHelpers";
 import {BarchartIndicator, Filter} from "../../../app/types";
 import AreaIndicatorsTable from "../../../app/components/plots/table/AreaIndicatorsTable.vue";
 import {switches} from "../../../app/featureSwitches";
+import ErrorAlert from "../../../app/components/ErrorAlert.vue";
 
 const localVue = createLocalVue();
 
-function getStore(modelOutputState: Partial<ModelOutputState> = {}, partialGetters = {}, partialSelections = {}, barchartFilters: any = ["TEST BAR FILTERS"], comparisonPlotFilters: any = ["TEST COMPARISON FILTERS"]) {
+function getStore(modelOutputState: Partial<ModelOutputState> = {}, partialGetters = {}, partialSelections = {}, barchartFilters: any = ["TEST BAR FILTERS"], comparisonPlotFilters: any = ["TEST COMPARISON FILTERS"], comparisonPlotError: any = null) {
     const store = new Vuex.Store({
         state: emptyState(),
         modules: {
@@ -46,7 +48,8 @@ function getStore(modelOutputState: Partial<ModelOutputState> = {}, partialGette
                 state: mockModelCalibrateState(
                     {
                         result: mockCalibrateResultResponse({data: ["TEST DATA"] as any}),
-                        comparisonPlotResult: mockComparisonPlotResponse({data: ["TEST COMPARISON DATA"] as any})
+                        comparisonPlotResult: mockComparisonPlotResponse({data: ["TEST COMPARISON DATA"] as any}),
+                        comparisonPlotError
                     }
                 )
             },
@@ -177,6 +180,14 @@ describe("ModelOutput component", () => {
         expect(comparisonPlot.props().formatFunction).toBe(vm.formatBarchartValue);
         expect(comparisonPlot.props().showRangesInTooltips).toBe(true);
         expect(comparisonPlot.props().disaggregateByConfig).toStrictEqual({fixed: true, hideFilter: true});
+    });
+
+    it("renders comparison plot error", () => {
+        const error = mockError("comparison plot error occurred")
+        const store = getStore({selectedTab: "comparison"}, {}, {}, [], [], error);
+        const wrapper = shallowMount(ModelOutput, {localVue, store});
+        expect(wrapper.findAll(ErrorAlert).length).toBe(1);
+        expect(wrapper.find(ErrorAlert).props().error).toBe(error);
     });
 
     it("does not render comparison plot if no there are no comparison plot indicators", () => {
@@ -713,7 +724,6 @@ describe("ModelOutput component", () => {
         expect(table.props().indicators).toStrictEqual(["TEST CHORO INDICATORS"]);
         expect(table.props().tableData).toStrictEqual(["TEST DATA"]);
         expect(table.props().countryAreaFilterOption).toStrictEqual({TEST: "TEST countryAreaFilterOption"});
-        expect(table.props().translateLabels).toBe(undefined);
     });
 
     it("renders choropleth table with correct indicator props", () => {
@@ -746,7 +756,6 @@ describe("ModelOutput component", () => {
         expect(table.props().indicators).toStrictEqual(["TEST BUBBLE INDICATORS", "TEST BUBBLE INDICATORS"]);
         expect(table.props().tableData).toStrictEqual(["TEST DATA"]);
         expect(table.props().countryAreaFilterOption).toStrictEqual({TEST: "TEST countryAreaFilterOption"});
-        expect(table.props().translateLabels).toBe(undefined);
     });
 
     it("renders bubble plot table with correct indicator props", () => {
@@ -805,7 +814,7 @@ describe("ModelOutput component", () => {
         });
         expect(table.props().tableData).toStrictEqual(["TEST DATA"]);
         expect(table.props().countryAreaFilterOption).toStrictEqual({TEST: "TEST countryAreaFilterOption"});
-        expect(table.props().translateLabels).toBe(undefined);
+        expect(table.props().translateFilterLabels).toBe(true);
     });
 
     it("renders barchart table with correct indicator props", () => {
@@ -823,6 +832,50 @@ describe("ModelOutput component", () => {
         const wrapper = shallowMount(ModelOutput, {localVue, store});
 
         const table = wrapper.find(AreaIndicatorsTable);
+        expect(table.props().selections).toStrictEqual({indicatorId: "art_coverage"});
+        expect(table.props().indicators).toStrictEqual(
+            [
+                {"indicator": "art_coverage", "indicator_value": "4"}
+            ]
+        );
+    });
+
+    it("renders comparison plot table", () => {
+        const store = getStore({selectedTab: "comparison"});
+        const wrapper = shallowMount(ModelOutput, {localVue, store});
+
+        const table = wrapper.findAll(AreaIndicatorsTable).at(1);
+        expect(table.props().areaFilterId).toBe("area");
+        expect(table.props().filters).toStrictEqual(["TEST COMPARISON FILTERS"]);
+        expect(table.props().selections).toStrictEqual({
+            indicatorId: "TestIndicator",
+            xAxisId: "age",
+            disaggregateById: "source",
+            selectedFilterOptions: {
+                region: [{id: "r1", label: "region 1"}],
+                age: [{id: "a1", label: "0-4"}]
+            }
+        });
+        expect(table.props().tableData).toStrictEqual(["TEST COMPARISON DATA"]);
+        expect(table.props().countryAreaFilterOption).toStrictEqual({TEST: "TEST countryAreaFilterOption"});
+        expect(table.props().translateFilterLabels).toBe(false);
+    });
+
+    it("renders comparison plot table with correct indicator props", () => {
+        const store = getStore({selectedTab: "comparison"}, {
+            comparisonPlotIndicators: jest.fn().mockReturnValue(
+                [
+                    {"indicator": "prevalence", "indicator_value": "2"},
+                    {"indicator": "art_coverage", "indicator_value": "4"}
+                ]
+            )
+        },
+            {
+                comparisonPlot: {indicatorId: "art_coverage"}
+            });
+        const wrapper = shallowMount(ModelOutput, {localVue, store});
+
+        const table = wrapper.findAll(AreaIndicatorsTable).at(1);
         expect(table.props().selections).toStrictEqual({indicatorId: "art_coverage"});
         expect(table.props().indicators).toStrictEqual(
             [