+++ src/app/static/src/tests/modelCalibrate/actions.test.ts
@@ -418,7 +418,7 @@ describe("ModelCalibrate actions", () => {
         expect(commit.mock.calls.length).toBe(2);
         expect(commit.mock.calls[0][0]).toStrictEqual("ComparisonPlotStarted");
         expect(commit.mock.calls[1][0]).toStrictEqual({
-            type: "SetError",
+            type: "SetComparisonPlotError",
             payload: mockError("Test Error")
         });
         expect(mockAxios.history.get.length).toBe(1);